Remove cureItem and cureAura logic, add migration notes
The logic in cureItem.lua and most of cureAura.lua has been removed and replaced with comments indicating migration to https://github.com/liantian-cn/RotationLib. The cure.toc file has been updated to comment out cureItem.lua and add a migration note. This change reflects the transition of functionality to the new library.